% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cluster_embeddings}
\alias{cluster_embeddings}
\title{Cluster embeddings using k-means}
\usage{
cluster_embeddings(x, n_clusters = 5, cluster_column = "cluster")
}
\arguments{
\item{x}{A tidyvec object}

\item{n_clusters}{Number of clusters}

\item{cluster_column}{Name for the cluster assignment column (default: "cluster")}
}
\value{
tidyvec object with cluster assignments added
}
\description{
Cluster embeddings using k-means
}
