% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{embedding_column}
\alias{embedding_column}
\title{Extract embedding column from a tidyvec object}
\usage{
embedding_column(x)
}
\arguments{
\item{x}{A tidyvec object}
}
\value{
Name of the embedding column
}
\description{
Extract embedding column from a tidyvec object
}
\keyword{internal}
