% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{embedding_fn}
\alias{embedding_fn}
\title{Extract embedding function from a tidyvec object}
\usage{
embedding_fn(x)
}
\arguments{
\item{x}{A tidyvec object}
}
\value{
The embedding function or NULL
}
\description{
Extract embedding function from a tidyvec object
}
\keyword{internal}
