% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{first_valid_embedding}
\alias{first_valid_embedding}
\title{Get first valid (non-NULL) embedding from a list}
\usage{
first_valid_embedding(emb_list)
}
\arguments{
\item{emb_list}{List of embeddings}
}
\value{
First non-NULL embedding or NULL if none found
}
\description{
Get first valid (non-NULL) embedding from a list
}
\keyword{internal}
