\name{\%~\%}
\alias{\%~\%}
\title{Similarity operator}
\usage{
a \%~\% b
}
\arguments{
\item{a}{First vector or tidyvec object}

\item{b}{Second vector or tidyvec object}

\item{method}{Similarity method to use ("cosine", "euclidean", or "dot")}
}
\value{
Similarity score or tidyvec object with similarity scores
}
\description{
Similarity operator
}
