% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{inspect_collection}
\alias{inspect_collection}
\title{Print details about a tidyvec collection}
\usage{
inspect_collection(x)
}
\arguments{
\item{x}{A tidyvec object}
}
\value{
Invisibly returns the input object
}
\description{
Print details about a tidyvec collection
}
