% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedders.R
\name{setup_python}
\alias{setup_python}
\title{Setup Python environment for TidyVec}
\usage{
setup_python(
  method = c("virtualenv", "conda"),
  envname = "tidyvec_env",
  packages = c("torch", "transformers", "pillow", "sentencepiece", "protobuf", "numpy")
)
}
\arguments{
\item{method}{Method for environment creation, either "virtualenv" or "conda"}

\item{envname}{Name of the environment to create}

\item{packages}{Additional packages to install}
}
\value{
Invisible TRUE if successful
}
\description{
Setup Python environment for TidyVec
}
