% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{vec}
\alias{vec}
\title{Creates a vector collection from a data frame or tibble}
\usage{
vec(x, embedding_column = "embedding", embedding_fn = NULL)
}
\arguments{
\item{x}{A data frame or tibble}

\item{embedding_column}{Name of column containing embeddings (or to be created)}

\item{embedding_fn}{Function to generate embeddings (optional)}
}
\value{
A tidyvec object
}
\description{
Creates a vector collection from a data frame or tibble
}
