% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write_vec}
\alias{write_vec}
\title{Save a tidyvec collection to disk}
\usage{
write_vec(x, file)
}
\arguments{
\item{x}{A tidyvec object}

\item{file}{Path to save file (recommended: .qs extension)}
}
\value{
Invisibly returns the input object
}
\description{
Save a tidyvec collection to disk
}
