% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedders.R
\name{embedder_hf}
\alias{embedder_hf}
\title{Create a HuggingFace embedding function}
\usage{
embedder_hf(
  model_name,
  modality = c("multimodal", "text", "image"),
  device = "cpu"
)
}
\arguments{
\item{model_name}{Name of the HuggingFace model}

\item{modality}{Type of model ("text", "image", or "multimodal")}

\item{device}{Device to use ("cpu", "cuda", or "mps")}
}
\value{
An embedding function
}
\description{
Create a HuggingFace embedding function
}
