% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{precision_at_k}
\alias{precision_at_k}
\title{Calculate Precision at k}
\usage{
precision_at_k(relevant_ids, retrieved_ids, k = 25)
}
\arguments{
\item{relevant_ids}{Character vector of IDs for relevant items}

\item{retrieved_ids}{Character vector of IDs for retrieved items}

\item{k}{Number of top results to consider (default: 25)}
}
\value{
Numeric value between 0 and 1
}
\description{
Computes the proportion of retrieved items that are relevant.
}
\details{
Precision@k = (# relevant items in top k) / k
}
\examples{
\dontrun{
relevant <- c("doc1", "doc2", "doc3")
retrieved <- c("doc1", "doc5", "doc2", "doc6", "doc3")
precision_at_k(relevant, retrieved, k = 5)  # 3/5 = 0.6
}

}
