% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{read_vec}
\alias{read_vec}
\title{Read a tidyvec collection from disk}
\usage{
read_vec(file)
}
\arguments{
\item{file}{Path to tidyvec collection file}
}
\value{
A tidyvec object
}
\description{
Read a tidyvec collection from disk
}
